Move user_interface_controller methods to controller and view classes
